Attempts make page run faster
1. Scrolling at 60 fps: In views/main.js, I updated the
updatePositions() function, so that rather than calling the document
body selector to get the top of the scroll for every pizza in the for
loop, it calls it outside and then uses that to calculate each pizza's
update.

2. Pizza resize: In main.js, I got rid of DetermineDx; rather than use
the pixel calculations, I used percents and just created an if-else
section that changes based on the update in the size.

3. Page Score (index.html): I minified files, made javascript async.